Fix missing incrementation in sider.list.List.__iter__()
This function have fetched a few items of the list that is too long
to be fetched at once. Furthermore, this code used the trick to calling
`Session.mark_query()` in the loop and prevent to calling it more than once.
But it doesn't need to do because `mark_query()` was called exactly one time
in the all situation. So I fixed it too.